LAjit- Accountancy Package
LAjit Technologies is a newly formed software development company,
delivering business solutions for all types of
industries. It offers a simplistic approach to the user experience with
"LAjit Solutions", while hiding the co mplexity and
strength of a corporate ERP system (FASB/GAAP compliant) behind the
scenes.Specialization:
• Payroll Software to
manage the Payroll of different companies catering to the entertainment
industries of US
• Works like an ERP system
taking the advantage of SOA• Independent of a specific platform or
technology.
• Customizable virtually for any industry
with any requirements without the intervention of the technical team
• Uses the state of the Art technology